\name{ccr.ExecuteMageck}
\alias{ccr.ExecuteMageck}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
    Executing MAGeCK from R command line
}
\description{
This function executes MAGeCK [1] from the command line, taking in input the path of the file containing the sgRNA counts' file
to be processed and saving the results in a user defined location. By default this function do not pre-normalise the counts. However this preliminary step can be included as specified by the corresponding argument. Additionally this function assumes that there is only one control sample, whose count values should be contained in the first column of the sgRNA counts' file.
This function requires python and the MAGeCK python package (v0.5.3, available at: \url{https://sourceforge.net/projects/mageck/files/0.5/mageck-0.5.3.zip/download}) to be installed. 
}
\usage{
ccr.ExecuteMageck(mgckInputFile,
                  expName = "expName",
                  normMethod = "none",
                  outputPath = "./")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{mgckInputFile}{
    A string specifying the path of the (plain text) file containing the sgRNA counts' file to be processed
}
  \item{expName}{
    A string specifying the experiment name. This is used as name prefix for all the files generated by MAGeCK.
}
  \item{normMethod}{
    A string specifying the normalisation method to be used ('none' by default).
}
  \item{outputPath}{
    A string specifying the folder where all the files outputted by MAGeCK will be saved.
}
}
\value{
    A string specifying the path to the gene summary file outputted by MAGeCK.
}
\references{
[1] Li, W., Xu, H., Xiao, T., Cong, L., Love, M. I., Zhang, F., et al. (2014). MAGeCK enables robust identification of essential genes from genome-scale CRISPR/Cas9 knockout screens. Genome Biology, 15(12), 554. [2] Hart, T., & Moffat, J. (2016). BAGEL: a computational framework for identifying essential genes from pooled library screens. BMC Bioinformatics, 17(1), 164.
}
\author{
Francesco Iorio (fi9323@gmail.com)
}
\examples{
    
## Loading sgRNA library annotation file
data(KY_Library_v1.0)

## Deriving the path of the file with the example dataset,
## from the mutagenesis of the EPLC-272H colorectal cancer cell line
fn<-paste(system.file('extdata', package = 'CRISPRcleanR'),
          '/EPLC-272H_counts.tsv',sep='')

## Loading, median-normalizing and computing fold-changes for the example dataset
normANDfcs<-ccr.NormfoldChanges(fn,min_reads=30,
                                EXPname='EPLC-272H',
                                libraryAnnotation = KY_Library_v1.0)

uncorrected_fn<-ccr.PlainTsvFile(sgRNA_count_object = normANDfcs$norm_counts,
                                 fprefix = 'EPLC-272H')

## execute MAGeCK saving files in the working directory
uncorrected_gs_fn<-ccr.ExecuteMageck(mgckInputFile = uncorrected_fn,
                                     expName = 'EPLC-272H',
                                     normMethod = 'none')
                                     
uncorrected_gs_fn
                                     
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{utils}% use one of  RShowDoc("KEYWORDS")
