\name{ccr.GWclean}
\alias{ccr.GWclean}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Unsupervised identification and correction of gene independent cell responses to CRISPR-Cas9 targeting.
}
\description{
 This function takes in input a genome-wide essentiality profile derived from a CRISPR-Cas9 experiment employing a pooled library of single guide RNAs (sgRNAs) targeting protein coding genes, which are transfected in an \emph{in vitro} model stably expressing Cas9.
    The essentiality profile quantifies the loss/gain-of-fitness caused by each sgRNA-targeting, and it is expressed as log fold changes (logFCs) between the
    aboundance of the sgRNAs at an end point after cell purification and their aboundance in the plasmid pool used for viral production, or at an initial time point, or in any other control condition. A circular binary segmentation algorithm [1, 2] is applied by this function to the genome-wide pattern of logFCs provided in input, in order to identify genomic regions including sgRNAs with sufficiently equal logFC (and mean logFC sufficiently different from background) and targeting a minimal number of different genes.
    Assuming that it is very unlikely to observe the same loss/gain-of-fitness effect when targeting a large number of contiguous genes, if certain user-defined condition (detailed below) are met then the logFCs of such regions are deemed as biased by some local feature of the involved genomic segment (which could be, for example, copy number amplified [3]), and they are corrected, i.e. mean centered [4]. \cr
}
\usage{
    ccr.GWclean(gwSortedFCs,label='',display=TRUE,
                saveTO=NULL,ignoredGenes=NULL,min.ngenes=3)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
 \item{gwSortedFCs}{
        A data frame containing genome-wide genomic-sorted sgRNAs' log fold changes. This data frame must include one named row per each sgRNA and the following columns/headers:
        \itemize{
        \item CHR: the chromosome of the gene targeted by the sgRNA under consideration;
        \item startp: the genomic coordinate of the starting position of the region targeted by the sgRNA under consideration;
        \item endp: the genomic coordinate of the ending position of the region targeted by the sgRNA under consideration;
        \item genes: the HGNC symbol of the gene targeted by the sgRNA under consideration;
        \item avgFC: the log fold change of the sgRNA under consideration averaged across replicates;
        \item BP: the genomic coordinate of the sgRNA defined as STARTpos+(ENDpos-STARTpos)/2.
        }
        This can be generated using the \code{\link{ccr.logFCs2chromPos}} function, starting from a data frame containing sgRNAs' log fold changes generated by the\cr \code{\link{ccr.NormfoldChanges}} function (from raw sgRNAs' counts), from raw sgRNAs' counts.
    }
    \item{label}{
        A string indicating the experiment name. This is used to compose the main title of the plots generated by this function and the name of the folder where the results are saved.
    }
    \item{display}{
        A logical value indicating whether genomic plots showing the results of the biased regions' identification and their log fold change correction should be generated or not.
    }
        \item{saveTO}{
        If different from NULL then this parameter will contain the path where pdf files with then genomic plots showing the results of the biased regions' identification (and their log fold change correction) will be saved (within a folder named as defined in the \code{label} parameter).   
        }
    \item{ignoredGenes}{
        A vector of strings containing HGNC symbols of genes that should not be considered when computing the minimal number of different genes targeted by sgRNAs in the same identified region of estimated equal log fold changes. This could contain, for example, a-priori known essential genes.
    }
    \item{min.ngenes}{
        A numerical value (>0) specifying the minimal number of different genes that the set of sgRNAs within a region of estimated equal logFCs should target in order for theri logFCs to be corrected, i.e. mean centered.
}
}

\value{
    A list containing two data frames and a vector of strings. The first data frame (corrected_logFCs) contains a named row per each sgRNA and the following columns/header:
        \itemize{
            \item \code{CHR}: the chromosome of the gene targeted by the sgRNA under consideration;
            \item \code{startp}: the genomic coordinate of the starting position of the region targeted by the sgRNA under consideration;
            \item \code{endp}: the genomic coordinate of the ending position of the region targeted by the sgRNA under consideration;
            \item \code{genes}: the HGNC symbol of the gene targeted by the sgRNA under consideration;
            \item \code{avgFC}: the log fold change of the sgRNA averaged across replicates;
            \item \code{correction}: the type of correction: 1 = increased log fold change, -1 = decreased log fold change. 0 indicates no correction;
            \item \code{correctedFC}: the corrected log fold change of the sgRNA
            }.\cr
    The second data frame (segments) contains the identified region of estimated equal log fold changes (one region per row) and the following
    columns/headers:
        \itemize{
            \item \code{CHR}: the chromosome of the gene targeted by the sgRNA under consideration;
            \item \code{startp}: the genomic coordinate of the starting position of the region targeted by the sgRNA under consideration;
            \item \code{endp}: the genomic coordinate of the ending position of the region targeted by the sgRNA under consideration;
            \item \code{genes}: the HGNC symbol of the gene targeted by the sgRNA under consideration;
            \item \code{avgFC}: the log fold change of the sgRNA averaged across replicates;
            \item \code{correction}: the type of correction: 1 = increased log fold change, -1 = decreased log fold change. 0 indicates no correction;
            \item \code{correctedFC}: the corrected log fold change of the sgRNA}.\cr
    The second data frame (segments) contains the identified region of estimated equal log fold changes (one region per row) and the following
    columns/headers:
        \itemize{
            \item \code{CHR}: the chromosome of the region under consideration;
            \item \code{startp}: the genomic coordinate of the starting position of the region under consideration;
            \item \code{endp}: the genomic coordinate of the ending position of the region under consideration;
            \item \code{n.sgRNAs}: the number of sgRNAs targeting sequences in the region under consideration;
            \item \code{avg.logFC}: the average log fold change of the sgRNAs in the region;
            \item \code{guideIdx}: the indexes range of the sgRNAs targeting the region under consideration as they appear in the gwSortedF Cs provided in input.
            }
    The string of vectors (SORTED_sgRNAs) contains the sgRNAs' identifiers in the same order as they are reported in the gwSortedFCs input data frame, i.e. genome sorted.
}
\references{
   [1] Olshen, A. B., Venkatraman, E. S., Lucito, R., Wigler, M. (2004). \emph{Circular binary segmentation for the analysis of array-based DNA copy number data}. Biostatistics 5: 557-572. \
    
    [2] Venkatraman, E. S., Olshen, A. B. (2007). \emph{A faster circular binary segmentation algorithm for the analysis of array CGH data}. Bioinformatics 23: 657-63. \
    
    [3] Andrew J. Aguirre, Robin M. Meyers, Barbara A. Weir, Francisca Vazquez, Cheng-Zhong Zhang, Uri Ben-David, April Cook, Gavin Ha, William F. Harrington, Mihir B. Doshi, Maria Kost-Alimova, Stanley Gill, Han Xu, Levi D. Ali, Guozhi Jiang, Sasha Pantel, Yenarae Lee, Amy Goodale, Andrew D. Cherniack, Coyin Oh, Gregory Kryukov, Glenn S. Cowley, Levi A. Garraway, Kimberly Stegmaier, Charles W. Roberts, Todd R. Golub, Matthew Meyerson, David E. Root, Aviad Tsherniak and William C. Hahn. \emph{Genomic copy number dictates a gene-independent cell response to CRISPR-Cas9 targeting}. Cancer Discov June 3 2016 DOI: 10.1158/2159-8290.CD-16-0154
    
    [4] Iorio, F., Behan, F. M., Goncalves, E., Beaver, C., Ansari, R., Pooley, R., et al. (n.d.). Unsupervised correction of gene-independent cell responses to CRISPR-Cas9 targeting. \cr
http://doi.org/10.1101/228189
}
\author{
Francesco Iorio (iorio@ebi.ac.uk)
}
\seealso{
\code{\link{ccr.cleanChrm}}
}
\examples{

## Loading sgRNA library annotation file
data(KY_Library_v1.0)

## Deriving the path of the file with the example dataset,
## from the mutagenesis of the HT-29 colorectal cancer cell line
fn<-paste(system.file('extdata', package = 'CRISPRcleanR'),'/HT-29_counts.tsv',sep='')

## Loading, median-normalizing and computing fold-changes for the example dataset
normANDfcs<-ccr.NormfoldChanges(fn,min_reads=30,EXPname='HT-29',
                                libraryAnnotation = KY_Library_v1.0)

## Genome-sorting of the fold changes
gwSortedFCs<-ccr.logFCs2chromPos(normANDfcs$logFCs,KY_Library_v1.0)

## Identifying and correcting biased sgRNAs' fold changes
correctedFCs<-ccr.GWclean(gwSortedFCs,display=TRUE,label='HT-29')

## Visualising first five entries of the corrected fold changes
head(correctedFCs$corrected_logFCs)

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{analysis}% use one of  RShowDoc("KEYWORDS")

