\name{ccr.sgRNAmeanFCs}
\alias{ccr.sgRNAmeanFCs}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Extract corrected logFC in vectorial format.
}
\description{
 This function takes as input a data frame containing the corrected logFC in the format generated by the \link{ccr.GWclean} function.
  This is a utility function that runs automatically as part of the \link{ccr.AnalysisPipeline}.
 \cr
}
\usage{
  ccr.sgRNAmeanFCs(
    foldchanges
  ) 
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{foldchanges}{
    A data frame with a "correctedFC" column storing the CN corrected logFC and row names equal to the sgRNA associated to the foldchages.
  }
}
\value{
  A named vector containing the corrected logFC and names equal to the related sgRNAs.
}
\author{
Paolo Cremaschi (paolo.crmeaschi@fht.org)
}
\seealso{
\code{
  \link{ccr.AnalysisPipeline}
  \link{ccr.GWclean}
}}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{analysis}% use one of  RShowDoc("KEYWORDS")

