\name{EssGenes.KEGG_rna_polymerase}
\alias{EssGenes.KEGG_rna_polymerase}
\docType{data}
\title{
 Core Fitness essential rna polymerase genes
}
\description{
List of core fitness essential rna polymerase genes downloaded from MSigDB [1].
}
\usage{data("EssGenes.KEGG_rna_polymerase")}
\format{
 A vector of strings containing HGNC symbols.
}
\references{
[1] Subramanian, A., Tamayo, P., Mootha, V. K., Mukherjee, S., Ebert, B. L., Gillette, M. A., et al. (2005). Gene set enrichment analysis: a knowledge-based approach for interpreting genome-wide expression profiles. Proceedings of the National Academy of Sciences of the United States of America, 102(43), 15545-15550. http://doi.org/10.1073/pnas.0506580102 \cr

[2] Iorio, F., Behan, F. M., Goncalves, E., Beaver, C., Ansari, R., Pooley, R., et al. (n.d.). Unsupervised correction of gene-independent cell responses to CRISPR-Cas9 targeting.\cr
http://doi.org/10.1101/228189
}
\examples{
data(EssGenes.KEGG_rna_polymerase)
head(EssGenes.KEGG_rna_polymerase)
}
\keyword{datasets}
