\name{EssGenes.ribosomalProteins}
\alias{EssGenes.ribosomalProteins}
\docType{data}
\title{
 Core Fitness essential genes coding for ribosomal proteins
}
\description{
List of core fitness essential coding for ribosomal proteins curated from [1].
}
\usage{data("EssGenes.KEGG_rna_polymerase")}
\format{
 A vector of strings containing HGNC symbols.
}
\references{
[1] Yoshihama, M. et al. The human ribosomal protein genes: sequencing and comparative analysis of 73 genes. Genome Res. 12, 379-390 (2002)

[2] Iorio, F., Behan, F. M., Goncalves, E., Beaver, C., Ansari, R., Pooley, R., et al. (n.d.). Unsupervised correction of gene-independent cell responses to CRISPR-Cas9 targeting. \cr
http://doi.org/10.1101/228189
}
\examples{
data(EssGenes.ribosomalProteins)
head(EssGenes.ribosomalProteins)
}
\keyword{datasets}
